requirements.txt
anthropic>=0.40.0
pydantic>=2.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
PyMuPDF>=1.24.0
tinydb>=4.8.0
python-docx>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
